pyproject.toml
[project]
name = "agent-development-kit"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-adk>=1.22.1",
    "litellm>=1.81.0",
    "mypy>=1.19.1",
    "ruff>=0.14.13",
]
